+++ controller/frontend/src/Controller/Frontend/Basket/Decorator/Category.php
@@ -25,7 +25,7 @@ class Category
 	 * Adds a product to the basket of the customer stored in the session
 	 *
 	 * @param \Aimeos\MShop\Product\Item\Iface $product Product to add including texts, media, prices, attributes, etc.
-	 * @param int $quantity Amount of products that should by added
+	 * @param float $quantity Amount of products that should by added
 	 * @param array $variant List of variant-building attribute IDs that identify an article in a selection product
 	 * @param array $config List of configurable attribute IDs the customer has chosen from
 	 * @param array $custom Associative list of attribute IDs as keys and arbitrary values that will be added to the ordered product
@@ -36,7 +36,7 @@ class Category
 	 * @throws \Aimeos\Controller\Frontend\Basket\Exception If the product isn't available
 	 */
 	public function addProduct( \Aimeos\MShop\Product\Item\Iface $product,
-		int $quantity = 1, array $variant = [], array $config = [], array $custom = [],
+		float $quantity = 1, array $variant = [], array $config = [], array $custom = [],
 		string $stocktype = 'default', string $supplier = '', string $siteid = null ) : \Aimeos\Controller\Frontend\Basket\Iface
 	{
 		$context = $this->getContext();
