+++ controller/frontend/src/Controller/Frontend/Basket/Standard.php
@@ -209,7 +209,7 @@ class Standard
 	 * Adds a product to the basket of the customer stored in the session
 	 *
 	 * @param \Aimeos\MShop\Product\Item\Iface $product Product to add including texts, media, prices, attributes, etc.
-	 * @param int $quantity Amount of products that should by added
+	 * @param float $quantity Amount of products that should by added
 	 * @param array $variant List of variant-building attribute IDs that identify an article in a selection product
 	 * @param array $config List of configurable attribute IDs the customer has chosen from
 	 * @param array $custom Associative list of attribute IDs as keys and arbitrary values that will be added to the ordered product
@@ -220,7 +220,7 @@ class Standard
 	 * @throws \Aimeos\Controller\Frontend\Basket\Exception If the product isn't available
 	 */
 	public function addProduct( \Aimeos\MShop\Product\Item\Iface $product,
-		int $quantity = 1, array $variant = [], array $config = [], array $custom = [],
+		float $quantity = 1, array $variant = [], array $config = [], array $custom = [],
 		string $stocktype = 'default', string $supplier = '', string $siteid = null ) : Iface
 	{
 		$this->checkAttributes( [$product], 'custom', array_keys( $custom ) );
@@ -274,10 +274,10 @@ class Standard
 	 * Edits the quantity of a product item in the basket.
 	 *
 	 * @param int $position Position number (key) of the order product item
-	 * @param int $quantity New quantiy of the product item
+	 * @param float $quantity New quantiy of the product item
 	 * @return \Aimeos\Controller\Frontend\Basket\Iface Basket frontend object for fluent interface
 	 */
-	public function updateProduct( int $position, int $quantity ) : Iface
+	public function updateProduct( int $position, float $quantity ) : Iface
 	{
 		$product = $this->get()->getProduct( $position );
